crm_bot-main/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-telegram-bot>=20.0
python-dotenv>=0.21.0
aiohttp>=3.8.0
matplotlib>=3.7.0
